# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Baby Health Application API
  version: 1.0.0
  description: Multi-baby health record API (India). Reference values are returned only with provenance and only when their release gate is CLEARED.
    The API never returns a computed milk target for healthy infants.
  contact:
    name: Product owner
servers:
- url: https://api.example.in
  description: Production (India region) - placeholder
security:
- bearerAuth: []
...
paths:
...
  /v1/vaccine-schedules:
    get:
      summary: List schedules
      tags:
      - Vaccinations
      x-authorization: Any authenticated user
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Schedule'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
...
  /v1/babies/{babyId}/schedule-selection:
    parameters:
    - $ref: '#/components/parameters/BabyId'
    get:
      summary: Current schedule selection
      tags:
      - Vaccinations
      x-authorization: Any active member of the baby (OWNER, GUARDIAN, CAREGIVER, VIEWER)
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleSelection'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
    put:
      summary: Select / switch schedule (versioned, audited)
      tags:
      - Vaccinations
      x-authorization: OWNER, GUARDIAN
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleSelection'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleSelection'
  /v1/babies/{babyId}/vaccinations/plan:
    parameters:
    - $ref: '#/components/parameters/BabyId'
    get:
      summary: Due / overdue / given plan
      tags:
      - Vaccinations
      x-authorization: Any active member of the baby (OWNER, GUARDIAN, CAREGIVER, VIEWER)
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlanItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '409':
          $ref: '#/components/responses/ScheduleNotReleased'
  /v1/babies/{babyId}/vaccinations:
    parameters:
    - $ref: '#/components/parameters/BabyId'
    get:
      summary: List vaccinations
      tags:
      - Vaccinations
      x-authorization: Any active member of the baby (OWNER, GUARDIAN, CAREGIVER, VIEWER)
      parameters:
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/From'
      - $ref: '#/components/parameters/To'
      - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Vaccination'
                  next_cursor:
                    type:
                    - string
                    - 'null'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
    post:
      summary: Create vaccinations record
      tags:
      - Vaccinations
      x-authorization: OWNER, GUARDIAN
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vaccination'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '409':
          $ref: '#/components/responses/DuplicateSuspected'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VaccinationCreate'
  /v1/babies/{babyId}/vaccinations/{id}:
    parameters:
    - $ref: '#/components/parameters/BabyId'
    - $ref: '#/components/parameters/Id'
    get:
      summary: Get vaccinations record
      tags:
      - Vaccinations
      x-authorization: Any active member of the baby (OWNER, GUARDIAN, CAREGIVER, VIEWER)
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vaccination'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
    patch:
      summary: Update vaccinations record
      tags:
      - Vaccinations
      x-authorization: OWNER, GUARDIAN
      parameters:
      - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vaccination'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '412':
          $ref: '#/components/responses/VersionMismatch'
        '422':
          $ref: '#/components/responses/RuleViolation'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VaccinationPatch'
    delete:
      summary: Soft-delete vaccinations record
      tags:
      - Vaccinations
      x-authorization: OWNER, GUARDIAN
      parameters:
      - $ref: '#/components/parameters/IfMatch'
      responses:
        '204':
          description: No Content
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '412':
          $ref: '#/components/responses/VersionMismatch'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token <= 15 min; authorization resolved per baby via membership + database row-level security.
  parameters:
    BabyId:
      name: babyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Resources of babies without membership return 404.
...
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    From:
      name: from
      in: query
      schema:
        type: string
        format: date-time
    To:
      name: to
      in: query
      schema:
        type: string
        format: date-time
    Sort:
      name: sort
      in: query
      schema:
        type: string
        enum:
        - occurred_at
        - -occurred_at
        - created_at
        - -created_at
        default: -occurred_at
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
      description: Record version
  responses:
    BadRequest:
      description: Validation failed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://errors.example/400
            title: Validation failed
            status: 400
    Unauthenticated:
      description: Authentication required
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://errors.example/401
            title: Authentication required
            status: 401
    Forbidden:
      description: Role does not permit this action
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://errors.example/403
            title: Role does not permit this action
            status: 403
    NotFound:
      description: Not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://errors.example/404
            title: Not found
            status: 404
    DuplicateSuspected:
      description: Possible duplicate; resend with force=true to save
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://errors.example/409
            title: Possible duplicate; resend with force=true to save
            status: 409
    ScheduleNotReleased:
      description: Schedule not released (verification pending)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://errors.example/409
            title: Schedule not released (verification pending)
            status: 409
...
    VersionMismatch:
      description: Record changed; refetch
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://errors.example/412
            title: Record changed; refetch
            status: 412
...
    RuleViolation:
      description: Business rule violated (e.g. quantity sent for DIRECT_BREASTFEEDING)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://errors.example/422
            title: Business rule violated (e.g. quantity sent for DIRECT_BREASTFEEDING)
            status: 422
    RateLimited:
      description: Too many requests (see Retry-After)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://errors.example/429
            title: Too many requests (see Retry-After)
            status: 429
    Internal:
      description: Internal error (support code provided)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://errors.example/500
            title: Internal error (support code provided)
            status: 500
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem details
      required:
      - type
      - title
      - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
        support_code:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              code:
                type: string
              message:
                type: string
    Audit:
      type: object
      properties:
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        created_by:
          type: string
          format: uuid
        version:
          type: integer
    Provenance:
      type: object
      required:
      - source_id
      - organization
      - document_title
      - source_url
      - page_or_section
      - verification_status
      properties:
        source_id:
          type: string
        organization:
          type: string
        document_title:
          type: string
        document_type:
          type: string
        country:
          type: string
        population:
          type: string
        clinical_context:
          type: string
        source_url:
          type: string
        publication_date:
          type: string
        version:
          type: string
        page_or_section:
          type: string
        retrieved_at:
          type: string
        verified_at:
          type: string
        evidence_level:
          type: string
        verification_status:
          type: string
        notes:
          type: string
...
    Schedule:
      type: object
      properties:
        schedule_id:
          type: string
        source:
          type: string
        country:
          type: string
        applicability:
          type: string
        versions:
          type: array
          items:
            type: object
    ScheduleSelection:
      type: object
      required:
      - schedule_id
      properties:
        schedule_id:
          enum:
          - GOVERNMENT_OF_INDIA_UIP
          - IAP_RECOMMENDED_SCHEDULE
        je_opt_in:
          type: boolean
    PlanItem:
      type: object
      properties:
        schedule_item_id:
          type: string
        vaccine:
          type: string
        dose_label:
          type: string
        due_from:
          type: string
          format: date
        due_to:
          type:
          - string
          - 'null'
          format: date
        status:
          enum:
          - GIVEN
          - DUE
          - OVERDUE
          - UPCOMING
          - NOT_APPLICABLE
          - CONFLICT_REQUIRES_REVIEW
        vaccination_id:
          type:
          - string
          - 'null'
        provenance:
          $ref: '#/components/schemas/Provenance'
    VaccinationCreate:
      type: object
      required:
      - vaccine_name_as_recorded
      - given_on
      properties:
        vaccine_id:
          type:
          - string
          - 'null'
        vaccine_name_as_recorded:
          type: string
        dose_label:
          type:
          - string
          - 'null'
        given_on:
          type: string
          format: date
        schedule_item_id:
          type:
          - string
          - 'null'
        clinic_id:
          type:
          - string
          - 'null'
          format: uuid
        given_by:
          type:
          - string
          - 'null'
        lot_number:
          type:
          - string
          - 'null'
        card_document_id:
          type:
          - string
          - 'null'
          format: uuid
        appointment_id:
          type:
          - string
          - 'null'
          format: uuid
        notes:
          type:
          - string
          - 'null'
        force:
          type: boolean
    VaccinationPatch:
      type: object
      properties:
        vaccine_id:
          type:
          - string
          - 'null'
        vaccine_name_as_recorded:
          type: string
        dose_label:
          type:
          - string
          - 'null'
        given_on:
          type: string
          format: date
        schedule_item_id:
          type:
          - string
          - 'null'
        clinic_id:
          type:
          - string
          - 'null'
          format: uuid
        given_by:
          type:
          - string
          - 'null'
        lot_number:
          type:
          - string
          - 'null'
        card_document_id:
          type:
          - string
          - 'null'
          format: uuid
        appointment_id:
          type:
          - string
          - 'null'
          format: uuid
        notes:
          type:
          - string
          - 'null'
        force:
          type: boolean
    Vaccination:
      allOf:
      - $ref: '#/components/schemas/VaccinationCreate'
      - $ref: '#/components/schemas/Audit'
      - type: object
        properties:
          id:
            type: string
            format: uuid
          schedule_version_id:
            type:
            - string
            - 'null'
...